requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
sentence-transformers
faiss-cpu